MAGECLOUD-2151: Add possibility to merge settings for Elastic search
